Fixed wall drawing, added FPS to UI
Previously, every wall corner (the image between four walls) was rotated
during the draw() method for the walls. Now, images are pre-rotated and
stored in the Resources.wCorners, and the array has been expanded in a
dimension to store them.
Overall, this greatly increases FPS while drawing walls, especially if
there are many of them or if the graphics are scaled. This helps the
FPS drop from ~50 to ~25 that occurred while flying / zooming out (FPS
is now more around 35-40).